Updated wiris plugin link
Also updated Ingress and Service object creation
to accommodate for when the Operator is being run
on Minikube.

Essentially when using on Minikube DomainName is not
needed to be specified in the Moodle Spec.
In this case the Ingress object does not need to have
TLS fields.
Similarly, making the Service object of type NodePort
instead of ClusterIP

Tagging a new release: 0.4.8